+++ src/Oro/Bundle/MagentoBundle/Tests/Functional/Fixtures/LoadGuestCustomerStrategyData.php
@@ -119,7 +119,6 @@ class LoadGuestCustomerStrategyData extends LoadMagentoChannel
         $customer->setGroup($this->customerGroup);
         // DateTimeZones should be removed in BAP-8710. Tests should be passed for:
         //  - Oro\Bundle\MagentoBundle\Tests\Functional\Controller\Api\Rest\CustomerControllerTest
-        //  - Oro\Bundle\MagentoBundle\Tests\Functional\Controller\Api\Rest\MagentoCustomerControllerTest
         $customer->setCreatedAt(new \DateTime('now', new \DateTimezone('UTC')));
         $customer->setUpdatedAt(new \DateTime('now', new \DateTimezone('UTC')));
         $customer->addAddress($address);